Refactored AgencyContainerV2 - moved the v2 dir out of agency v1 dir - Separated container with overview data request from page UI logic - moved AgencySection to its own file
