+++ src/js/containers/agencyV2/accountSpending/CountTabContainer.jsx
@@ -6,7 +6,7 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { fetchSpendingCount } from 'helpers/agencyV2Helper';
-import CountTab from 'components/agency/v2/CountTab';
+import CountTab from 'components/agencyV2/CountTab';
 
 const propTypes = {
     fy: PropTypes.string.isRequired,
